requirements.txt
pandas
numpy
scikit-learn
click
fastapi
uvicorn
SQLAlchemy
PyYAML
requests